Create README.md
At the most basic level, a Chatbot is a computer program that simulates and
processes human conversation (either in text or speech), allowing humans to interact with
digital devices as if they are communicating with a real person. Chatbots work on the
principle of AI (Artificial Intelligence) and ML (Machine Learning). We, therefore,
shall be implementing a virtual assistant based on these principles which can solve any
college related query. This will work as a College Oriented Intelligence machine. This virtual
machine will respond to the queries of students as well as visitors on college related issues
.This chatbot will help students,teachers and especially new visitors to get any information
regarding the college quickly.
The college website may contain all the details required to know about the college, but this
chatbot shall provide an easy and convenient access to all those information.